fix bug
处理了一个小小的bug

+++ src/codecleanwithcreatmethod/RefactoringMortgagePaymentsCalculator.java
@@ -43,7 +43,7 @@ public class RefactoringMortgagePaymentsCalculator {
         double value;
         while (true) {
             System.out.print(prompt);
-            value = scanner.nextInt();
+            value = scanner.nextDouble();
             if (value >= min && value <= max)
                 break;
             System.out.println("Enter a value between " +min + " and "+ max);
